+++ precompiles/liquidstake/tx.go
@@ -2,6 +2,7 @@ package liquidstake
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
@@ -12,6 +13,11 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	keeper "github.com/cosmos/evm/x/liquidstake/keeper"
 	types "github.com/cosmos/evm/x/liquidstake/types"
+
+	evmtypes "github.com/cosmos/evm/x/vm/types"
+	cmn "github.com/cosmos/evm/precompiles/common"
+	"github.com/cosmos/evm/utils"
+
 )
 
 const (
@@ -36,6 +42,46 @@ const (
 	StakeToLPAuthz = types.AuthorizationType_AUTHORIZATION_TYPE_STAKE_TO_LP
 )
 
+//	var (
+//		// stakeAuthz is the authorization grant for the caller and the delegator address
+//		stakeAuthz *stakingtypes.StakeAuthorization
+//		// expiration is the expiration time of the authorization grant
+//		expiration *time.Time
+//
+//		// isCallerOrigin is true when the delegator is the same as the origin and is the caller address
+//		isCallerOrigin = delegatorHexAddr == origin && contract.CallerAddress == origin
+//		// isSCDelegator is true when the contract caller is the same as the delegator
+//		// and is not origin (it is a SmartContract)
+//		isSCDelegator = contract.CallerAddress == delegatorHexAddr && origin != contract.CallerAddress
+//	)
+//
+//	// 3 possible cases:
+//	// 1. Delegator is EOA and submits tx to stake its own funds (origin == contract_caller_addr) -> no auth needed
+//	// 2. Delegator is SC and submits tx to stake its own funds -> no auth needed (should be handled at SC level)
+//	// 3. Delegator is EOA and SC makes call to stake the EOA's funds -> auth needed
+//
+//	// no need to have authorization when the delegator is the owner of the funds
+//	if !isCallerOrigin && !isSCDelegator {
+//		// Check if the authorization grant exists for the caller and the origin
+//		stakeAuthz, expiration, err = authorization.CheckAuthzAndAllowanceForGranter(ctx, p.AuthzKeeper, contract.CallerAddress, delegatorHexAddr, &msg.Amount, DelegateMsg)
+//		if err != nil {
+//			return nil, err
+//		}
+//	}
+//
+//	// Execute the transaction using the message server
+//	msgSrv := stakingkeeper.NewMsgServerImpl(&p.stakingKeeper)
+//	if _, err = msgSrv.Delegate(ctx, msg); err != nil {
+//		return nil, err
+//	}
+//
+//	// Only update the authorization if the contract caller is different from owner of the funds
+//	if !isCallerOrigin && !isSCDelegator {
+//		if err := p.UpdateStakingAuthorization(ctx, contract.CallerAddress, delegatorHexAddr, stakeAuthz, expiration, DelegateMsg, msg); err != nil {
+//			return nil, err
+//		}
+//	}
+
 func (p Precompile) LiquidStake(
 	ctx sdk.Context,
 	origin common.Address,
@@ -50,30 +96,35 @@ func (p Precompile) LiquidStake(
 		return nil, err
 	}
 
-	msg, err := NewMsgLiquidStake(args, bondDenom)
+	delegatorHexAddr, msg, err := NewMsgLiquidStake(args, bondDenom)
 	if err != nil {
 		return nil, err
 	}
 
-	isCallerOrigin := contract.CallerAddress == origin
+	var (
+		// stakeAuthz is the authorization grant for the caller and the delegator address
+		liquidAuthz *types.LiquidStakeAuthorization
+		// expiration is the expiration time of the authorization grant
+		expiration *time.Time
 
-	delegatorAddr, err := sdk.AccAddressFromBech32(msg.DelegatorAddress)
-	if err != nil {
-		return nil, err
-	}
+		// isCallerOrigin is true when the delegator is the same as the origin and is the caller address
+		isCallerOrigin = *delegatorHexAddr == origin && contract.CallerAddress == origin
+		// isSCDelegator is true when the contract caller is the same as the delegator
+		// and is not origin (it is a SmartContract)
+		isSCDelegator = contract.CallerAddress == *delegatorHexAddr && origin != contract.CallerAddress
+	)
 
-	if !isCallerOrigin {
-		authz, expiration, err := authorization.CheckAuthzAndAllowanceForLiquidStake(ctx, p.AuthzKeeper, contract.CallerAddress, common.BytesToAddress(delegatorAddr), &msg.Amount, LiquidStakeMsg)
+	// 3 possible cases:
+	// 1. Delegator is EOA and submits tx to stake its own funds (origin == contract_caller_addr) -> no auth needed
+	// 2. Delegator is SC and submits tx to stake its own funds -> no auth needed (should be handled at SC level)
+	// 3. Delegator is EOA and SC makes call to stake the EOA's funds -> auth needed
+
+	// no need to have authorization when the delegator is the owner of the funds
+	if !isCallerOrigin && !isSCDelegator {
+		// Check if the authorization grant exists for the caller and the origin
+
+		liquidAuthz, expiration, err = authorization.CheckAuthzAndAllowanceForLiquidStake(ctx, p.AuthzKeeper, contract.CallerAddress, *delegatorHexAddr, &msg.Amount, LiquidStakeMsg)
 		if err != nil {
-			return nil, err
-		}
-
-		liquidAuthz, ok := authz.(*types.LiquidStakeAuthorization)
-		if !ok {
-			return nil, fmt.Errorf("unexpected authorization type: %T", authz)
-		}
-
-		if err := p.UpdateLiquidStakeAuthorization(ctx, contract.CallerAddress, common.BytesToAddress(delegatorAddr), liquidAuthz, expiration, LiquidStakeMsg, msg); err != nil {
 			return nil, err
 		}
 	}
@@ -83,6 +134,30 @@ func (p Precompile) LiquidStake(
 	// Execute the transaction using the message server
 	if _, err = msgSrv.LiquidStake(ctx, msg); err != nil {
 		return nil, err
+	}
+
+	// Only update the authorization if the contract caller is different from owner of the funds
+	if !isCallerOrigin && !isSCDelegator {
+		if err := p.UpdateLiquidStakeAuthorization(ctx, contract.CallerAddress, *delegatorHexAddr, liquidAuthz, expiration, LiquidStakeMsg, msg); err != nil {
+			return nil, err
+		}
+	}
+
+	if !isCallerOrigin && msg.Amount.Denom == evmtypes.GetEVMCoinDenom() {
+		// get the delegator address from the message
+		delAccAddr := sdk.MustAccAddressFromBech32(msg.DelegatorAddress)
+		delHexAddr := common.BytesToAddress(delAccAddr)
+		// NOTE: This ensures that the changes in the bank keeper are correctly mirrored to the EVM stateDB
+		// when calling the precompile from a smart contract
+		// This prevents the stateDB from overwriting the changed balance in the bank keeper when committing the EVM state.
+
+		// Need to scale the amount to 18 decimals for the EVM balance change entry
+		scaledAmt, err := utils.Uint256FromBigInt(evmtypes.ConvertAmountTo18DecimalsBigInt(msg.Amount.Amount.BigInt()))
+		if err != nil {
+			return nil, err
+		}
+
+		p.SetBalanceChangeEntries(cmn.NewBalanceChangeEntry(delHexAddr, scaledAmt, cmn.Sub))
 	}
 
 	return method.Outputs.Pack(true)
@@ -102,38 +177,26 @@ func (p Precompile) StakeToLP(
 		return nil, err
 	}
 
-	msg, err := NewMsgStakeToLP(args, liquidBondDenom, bondDenom)
+	delegatorHexAddr, msg, err := NewMsgStakeToLP(args, liquidBondDenom, bondDenom)
 	if err != nil {
 		return nil, err
 	}
 
-	isCallerOrigin := contract.CallerAddress == origin
+	var (
+		// isCallerOrigin is true when the delegator is the same as the origin and is the caller address
+		isCallerOrigin = *delegatorHexAddr == origin && contract.CallerAddress == origin
+		// isSCDelegator is true when the contract caller is the same as the delegator
+		// and is not origin (it is a SmartContract)
+		isSCDelegator = contract.CallerAddress == *delegatorHexAddr && origin != contract.CallerAddress
+	)
 
-//	delegatorAddr, err := sdk.AccAddressFromBech32(msg.DelegatorAddress)
-//	if err != nil {
-//		return nil, err
-//	}
+	// 2 possible cases:
+	// 1. Delegator is EOA and submits tx to stake its own funds (origin == contract_caller_addr) -> no auth needed
+	// 2. Delegator is SC and submits tx to stake its own funds -> no auth needed (should be handled at SC level)
 
-	//TODO: fix this
-	if !isCallerOrigin {
-		return nil, fmt.Errorf("TODO")
+	if !isCallerOrigin && !isSCDelegator {
+		return nil, fmt.Errorf("Delegator is not Origin nor caller, Delegation through precompile for StakeToLp is not possible")
 	}
-
-//	if !isCallerOrigin {
-//		authz, expiration, err := authorization.CheckAuthzAndAllowanceForLiquidStake(ctx, p.AuthzKeeper, contract.CallerAddress, common.BytesToAddress(delegatorAddr), &msg.StakedAmount, StakeToLPMsg)
-//		if err != nil {
-//			return nil, err
-//		}
-//
-//		liquidAuthz, ok := authz.(*types.LiquidStakeAuthorization)
-//		if !ok {
-//			return nil, fmt.Errorf("unexpected authorization type: %T", authz)
-//		}
-//
-//		if err := p.UpdateLiquidStakeAuthorization(ctx, contract.CallerAddress, common.BytesToAddress(delegatorAddr), liquidAuthz, expiration, StakeToLPMsg, msg); err != nil {
-//			return nil, err
-//		}
-//	}
 
 	msgSrv := keeper.NewMsgServerImpl(p.liquidStakeKeeper)
 
@@ -155,30 +218,35 @@ func (p Precompile) LiquidUnstake(
 ) ([]byte, error) {
 	bondDenom := p.liquidStakeKeeper.LiquidBondDenom(ctx)
 
-	msg, err := NewMsgLiquidUnstake(args, bondDenom)
+	delegatorHexAddr, msg, err := NewMsgLiquidUnstake(args, bondDenom)
 	if err != nil {
 		return nil, err
 	}
 
-	isCallerOrigin := contract.CallerAddress == origin
+	var (
+		// stakeAuthz is the authorization grant for the caller and the delegator address
+		liquidAuthz *types.LiquidStakeAuthorization
+		// expiration is the expiration time of the authorization grant
+		expiration *time.Time
 
-	delegatorAddr, err := sdk.AccAddressFromBech32(msg.DelegatorAddress)
-	if err != nil {
-		return nil, err
-	}
+		// isCallerOrigin is true when the delegator is the same as the origin and is the caller address
+		isCallerOrigin = *delegatorHexAddr == origin && contract.CallerAddress == origin
+		// isSCDelegator is true when the contract caller is the same as the delegator
+		// and is not origin (it is a SmartContract)
+		isSCDelegator = contract.CallerAddress == *delegatorHexAddr && origin != contract.CallerAddress
+	)
 
-	if !isCallerOrigin {
-		authz, expiration, err := authorization.CheckAuthzAndAllowanceForLiquidStake(ctx, p.AuthzKeeper, contract.CallerAddress, common.BytesToAddress(delegatorAddr), &msg.Amount, LiquidUnstakeMsg)
+	// 3 possible cases:
+	// 1. Delegator is EOA and submits tx to stake its own funds (origin == contract_caller_addr) -> no auth needed
+	// 2. Delegator is SC and submits tx to stake its own funds -> no auth needed (should be handled at SC level)
+	// 3. Delegator is EOA and SC makes call to stake the EOA's funds -> auth needed
+
+	// no need to have authorization when the delegator is the owner of the funds
+	if !isCallerOrigin && !isSCDelegator {
+		// Check if the authorization grant exists for the caller and the origin
+
+		liquidAuthz, expiration, err = authorization.CheckAuthzAndAllowanceForLiquidStake(ctx, p.AuthzKeeper, contract.CallerAddress, *delegatorHexAddr, &msg.Amount, LiquidUnstakeMsg)
 		if err != nil {
-			return nil, err
-		}
-
-		liquidAuthz, ok := authz.(*types.LiquidStakeAuthorization)
-		if !ok {
-			return nil, fmt.Errorf("unexpected authorization type: %T", authz)
-		}
-
-		if err := p.UpdateLiquidStakeAuthorization(ctx, contract.CallerAddress, common.BytesToAddress(delegatorAddr), liquidAuthz, expiration, LiquidUnstakeMsg, msg); err != nil {
 			return nil, err
 		}
 	}
@@ -190,6 +258,31 @@ func (p Precompile) LiquidUnstake(
 	if err != nil {
 		return nil, err
 	}
+
+	// Only update the authorization if the contract caller is different from owner of the funds
+	if !isCallerOrigin && !isSCDelegator {
+		if err := p.UpdateLiquidStakeAuthorization(ctx, contract.CallerAddress, *delegatorHexAddr, liquidAuthz, expiration, LiquidUnstakeMsg, msg); err != nil {
+			return nil, err
+		}
+	}
+
+	if !isCallerOrigin && msg.Amount.Denom == evmtypes.GetEVMCoinDenom() {
+		// get the delegator address from the message
+		delAccAddr := sdk.MustAccAddressFromBech32(msg.DelegatorAddress)
+		delHexAddr := common.BytesToAddress(delAccAddr)
+		// NOTE: This ensures that the changes in the bank keeper are correctly mirrored to the EVM stateDB
+		// when calling the precompile from a smart contract
+		// This prevents the stateDB from overwriting the changed balance in the bank keeper when committing the EVM state.
+
+		// Need to scale the amount to 18 decimals for the EVM balance change entry
+		scaledAmt, err := utils.Uint256FromBigInt(evmtypes.ConvertAmountTo18DecimalsBigInt(msg.Amount.Amount.BigInt()))
+		if err != nil {
+			return nil, err
+		}
+
+		p.SetBalanceChangeEntries(cmn.NewBalanceChangeEntry(delHexAddr, scaledAmt, cmn.Sub))
+	}
+
 
 	return method.Outputs.Pack(responce.CompletionTime.Unix())
 }
